+++ franka_mobile_sensors/launch/__init__.py
@@ -1,4 +1,4 @@
-#  Copyright (c) 2025 Franka Robotics GmbH
+#  Copyright (c) 2026 Franka Robotics GmbH
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -11,7 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from .config_loader import load_yaml_config
-
-__all__ = ['load_yaml_config']